refactor: Change the unused code and merge from staging
