+++ src/interface/return-interface.ts
@@ -1,0 +1,5 @@
+export interface FileUploadReturnInterface {
+  cloudStoragePublicUrl: string;
+  cloudStorageError: string;
+  storageObject: string;
+}
