+++ src/service/user/auth/register-service.ts
@@ -6,10 +6,15 @@ import { validate } from '../../../validation/validation';
 import bcrypt from 'bcryptjs';
 import { getUserByEmail } from '../../../utils/user/getUserByEmail';
 import { userCollection } from '../../../app/firestore';
+import { imageValidation } from '../../../validation/image/image-validation';
 
 const registerUser = async (request: RequestSignupInterface) => {
   const validatedUser = validate(registerUserValidation, request);
-  if (validatedUser) {
+  let validatedImage: any;
+  if (request.image) {
+    validatedImage = validate(imageValidation, request.image);
+  }
+  if (validatedUser || validatedImage) {
     const { userData } = await getUserByEmail(validatedUser.email);
 
     if (userData) {
