+++ src/validation/prediction/spice-classification-validation.ts
